removed redefinitions of FileDeleter
