+++ tests/timer_tests.cc
@@ -8,21 +8,9 @@
 #include "timer.h"
 #include "sqlite_back.h"
 
+#include "tools.h"
+
 static std::string const path = "testdb.sqlite";
-
-class FileDeleter {
- public:
-  FileDeleter(std::string path) {
-    path_ = path;
-  }
-
-  ~FileDeleter() {
-    remove(path_.c_str());
-  }
-
- private:
-  std::string path_;
-};
 
 class Dier : public cyclus::Facility {
  public:
